FR_722_integration_tests
- renamed folder name and configure build.gradle
- removed sonar exclusions
